Separate fractal renderer from GUI into distinct modules
Extract FractalRenderer into fractal_renderer.rs with no egui dependency.
Move EguiRenderer into ui.rs alongside UiState. Reduce renderer.rs to a
thin winit ApplicationHandler that orchestrates the two renderers.
